Refs #3612 : [BUG] [Endpoints] Always grant access if security is disabed
Added a property "endpointSecurityEnabled" to the ParamBean (meveo-admin.properties file) that indicates whether we want to secure the access to the endpoints.

I also fixed a bug that occured when we define a setter that is not a string and that we call an endpoint with GET, causing all the parameters to be gathered as string so we needed to convert them in the right type.

Signed-off-by: clement.bareth <clement.bareth@web-drone.fr>