Refs #3853 : Possibility to delete element referenced by module items
Before, uninstalling a module was making all the referenced elements to be disabled.
Now they can optionally be deleted.

Moreover, we were checking if the entity was making part of an other module regardless whether that module was active or not.
Now, if the module is not active, we do not skip the process of the entity.

In the GUI, the confirmation dialog now offers the option to remove the module items instead of disabling them.
In the API, a query parameter "remove" was added to the endpoint /module/uninstall, if set to true the module itemrs will be deleted. Default value to false.

Made remove(id) function from PersistenceService final so that all sub-classes use remove(entity) function.

Code clean-up and license update.

Signed-off-by: Clément Bareth <clement.bareth@gmail.com>